discord/pyproject.toml
[project]
name = "cavepedia-discord"
version = "0.1.0"
description = "Discord bot for Cavepedia AI assistant"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-json-logger>=4.0.0",
    "cachetools>=5.5.0",
]
